move auth from services to controllers
