+++ routes/auth.js
@@ -2,10 +2,12 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const router = express.Router();
 
-const authService = require('../services/auth');
+const controller = require('../controllers/auth');
 
 router.use(bodyParser.json());
 
-router.post('/login', authService.login);
+router.post('/login', async (req, res) => {
+    (await controller.login(req))(res);
+});
 
 module.exports = router;
