+++ scripts/createuser.js
@@ -2,7 +2,7 @@ const envLoader = require('dotenv-json');
 envLoader({ path: ".env.development.json"});
 
 const db = require('../services/database');
-const auth = require('../services/auth')
+const auth = require('../controllers/auth');
 
 let args = process.argv;
